refactor: add historical data tab
