+++ src/components/managedVaults/community/vaultDetails/VaultHistoricalData.tsx
@@ -1,0 +1,7 @@
+export default function VaultHistoricalData() {
+  return (
+    <div>
+      <h1>Vault Historical Data</h1>
+    </div>
+  )
+}
